Refactor edit-wrapper; Add edit guidelines
