+++ app/Livewire/Guidelines/ShowGuideline.php
@@ -7,46 +7,22 @@ use App\Models\Category;
 use App\Models\Criterion;
 use App\Models\Guideline;
 use Livewire\Attributes\Computed;
+use Livewire\Attributes\On;
 use Livewire\Component;
 
 class ShowGuideline extends Component
 {
     public Guideline $guideline;
-    public GuidelineForm $form;
 
-    public function mount(Guideline $guideline)
+    public function mount(Guideline $guideline): void
     {
         $this->guideline = $guideline;
-        $this->form->setModel($guideline);
     }
 
-    public function save(): void
+    #[On('version-updated')]
+    public function getGuideline(): void
     {
-        $this->authorize('update', $this->form->guideline);
-
-        $this->guideline = $this->form->update();
-
-        $this->dispatch('close-edit');
-    }
-
-    #[Computed] public function criterionOptions(): array
-    {
-        return Criterion::all()
-            ->map(fn ($criterion) => [
-                'value' => $criterion->id,
-                'option' => $criterion->getLongName(),
-            ])
-            ->toArray();
-    }
-
-    #[Computed] public function categoryOptions(): array
-    {
-        return Category::all()
-            ->map(fn ($category) => [
-                'value' => $category->id,
-                'option' => $category->name,
-            ])
-            ->toArray();
+        $this->guideline->refresh();
     }
 
     public function render()
