+++ resources/views/livewire/documents/show-document.blade.php
@@ -1,15 +1,16 @@
-<div x-data="{ showView: true }"
-     x-on:show-edit="showView = false"
-     x-on:close-edit="showView = true"
->
-    @can('update', $this->getDocument())
-        <livewire:documents.edit-document :slug="$slug" />
-    @endcan
+<div>
+    <x-forms.edit-wrapper>
+        <x-slot:view>
+            @if($title)
+                <h1>{{ $title }}</h1>
+            @endif
+            <div>{!! $content !!}</div>
+        </x-slot:view>
 
-    <div x-show="showView">
-        @if($title)
-            <h1>{{ $title }}</h1>
-        @endif
-        <div>{!! $content !!}</div>
-    </div>
+        @can('update', $this->getDocument())
+            <x-slot:edit>
+                <livewire:documents.edit-document :slug="$slug" />
+            </x-slot:edit>
+        @endcan
+    </x-forms.edit-wrapper>
 </div>
