+++ resources/views/livewire/guidelines/show-guideline.blade.php
@@ -13,7 +13,7 @@
         </h1>
     </div>
 
-    <x-forms.edit-model-wrapper :model="$guideline">
+    <x-forms.edit-wrapper>
         <x-slot:view>
             <h2>
                 {{ $guideline->name }}
@@ -49,20 +49,12 @@
             {!! Str::markdown($guideline->notes) !!}
         </x-slot:view>
 
-        <x-slot:edit>
-            <form wire:submit="save">
-                <x-forms.input label="Name" wire:model="form.name" />
-
-                <x-forms.select label="WCAG 2 criterion" wire:model="form.criterion_id" :options="$this->criterionOptions" />
-
-                <x-forms.select label="Category" wire:model="form.category_id" :options="$this->categoryOptions" />
-
-                <x-forms.textarea label="Notes" wire:model="form.notes" size="lg"/>
-
-                <x-forms.button.submit-group submitName="Update Guideline" />
-            </form>
-        </x-slot:edit>
-    </x-forms.edit-model-wrapper>
+        @can('update', $guideline)
+            <x-slot:edit>
+                <livewire:guidelines.edit-guideline :$guideline />
+            </x-slot:edit>
+        @endcan
+    </x-forms.edit-wrapper>
 
 </div>
